refactor: extract SyncUpdateDialog types to separate file
